requirements.txt
crewai
langchain
pyyaml
pandas
duckduckgo-search
python-dotenv
openai